moved scripts to separate folder
