+++ scripts/database/migration.ts
@@ -1,5 +1,5 @@
 import { migrate } from 'drizzle-orm/libsql/migrator';
-import { db } from './client';
+import { db } from '../../server/database/client';
 
 (async () => {
     await migrate(db, { migrationsFolder: './server/database/migrations' });
